Add support for BigInt columns
