+++ CBDatabase.podspec
@@ -16,4 +16,5 @@ Pod::Spec.new do |s|
 
   s.dependency 'RxSwift', '~> 4.4.0'
   s.dependency 'RxCocoa', '~> 4.4.0'
+  s.dependency 'BigInt', '~> 3.1'
 end
