+++ CBDatabase.xcworkspace/xcuserdata/hichambouabdallah.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -28,7 +28,7 @@
             ignoreCount = "0"
             continueAfterRunningActions = "No"
             filePath = "DatabasesTests/DatabasesTests.swift"
-            timestampString = "568788495.18325"
+            timestampString = "570055307.903523"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
             startingLineNumber = "38"
@@ -44,7 +44,7 @@
             ignoreCount = "0"
             continueAfterRunningActions = "No"
             filePath = "DatabasesTests/DatabasesTests.swift"
-            timestampString = "568788495.1835181"
+            timestampString = "570055307.9036191"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
             startingLineNumber = "20"
@@ -60,7 +60,7 @@
             ignoreCount = "0"
             continueAfterRunningActions = "No"
             filePath = "DatabasesTests/DatabasesTests.swift"
-            timestampString = "568788495.1836931"
+            timestampString = "570055307.903669"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
             startingLineNumber = "14"
