+++ DatabasesTests/TestDatabase.xcdatamodeld/v2.0.xcdatamodel/contents
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="14460.32" systemVersion="17G3025" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
+<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="14460.32" systemVersion="18C54" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
     <entity name="AdvancedModelCoreData" representedClassName="AdvancedModelCoreData" syncable="YES" codeGenerationType="class">
         <attribute name="customIdField" optional="YES" attributeType="String" syncable="YES"/>
     </entity>
@@ -17,7 +17,7 @@
         </uniquenessConstraints>
     </entity>
     <entity name="TestWallet" representedClassName="" syncable="YES">
-        <attribute name="balance" optional="YES" attributeType="Transformable" valueTransformerName="" customClassName="CBDatabaseTests.TestBigIntDBWrapper" syncable="YES"/>
+        <attribute name="balance" optional="YES" attributeType="Transformable" valueTransformerName="" customClassName="CBDatabase.BigIntDBWrapper" syncable="YES"/>
         <attribute name="id" optional="YES" attributeType="String" syncable="YES"/>
         <attribute name="name" optional="YES" attributeType="String" syncable="YES"/>
     </entity>
