+++ Podfile
@@ -5,13 +5,13 @@ target 'CBDatabase' do
 
   pod 'RxSwift', '~> 4.4.0'
   pod 'RxCocoa', '~> 4.4.0'
+  pod 'BigInt', '~> 3.1'
 
   target 'CBDatabaseTests' do
     inherit! :search_paths
     pod 'Quick', '~> 1.3.2'
     pod 'Nimble', '~> 7.3.1'
     pod 'RxBlocking', '~> 4.3.1'
-    pod 'BigInt', '~> 3.1'
   end
 
 end
